+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Updated Autosize Code Editor toolbar (#506)
+- Added custom input for Select, Multi select (#507)
 
 ## 4.6.0 (2024-09-28)